Cleaning up after merge.
