+++ firmware/src/MightyBoard/Motherboard/Command.cc
@@ -363,8 +363,8 @@ void runCommandSlice() {
 			} else if (command == HOST_CMD_CHANGE_TOOL) {
 				if (command_buffer.getLength() >= 2) {
 					command_buffer.pop(); // remove the command code
-                    currentToolIndex = command_buffer.pop();
-                    steppers::changeToolIndex(currentToolIndex);
+					currentToolIndex = command_buffer.pop();
+					planner::changeToolIndex(currentToolIndex);
 				}
 			} else if (command == HOST_CMD_ENABLE_AXES) {
 				if (command_buffer.getLength() >= 2) {
